requirements.txt
praw
nltk
numpy
torch
pandas
scikit-learn
python-dotenv
pytest